audio youtube and video tiktok. v1

+++ service/audioProcessing/audioDownloader.js
@@ -1,8 +1,9 @@
 const unorm = require("unorm")
-const { sendAudioFromFileId, downloadYoutubedl } = require("./audioFunction.js")
-const { getCleanVideoUrl } = require("../../utils/checkUrl.js")
-const { AudioFile } = require("../../models.js")
 const youtubedl = require("youtube-dl-exec")
+
+const { sendAudioFromFileId, downloadYoutubedl } = require("./audioFunction.js")
+const { checkYoutubeVideoUrl } = require("../../utils/checkUrl.js")
+const { AudioFile } = require("../../models.js")
 
 
 
@@ -17,64 +18,56 @@ const flags = {
 async function audioDownloader(bot, chatId, botName, audioDownloaderScene) {
 
   audioDownloaderScene.hears(/.*/, async (ctx) => {
+    
     try {
-      let normalizedFilename
-      let audioFile
-      let videoTitle
-      // проверяем, что сообщение от того же пользователя, которому отправили запрос
-      // if (text === "/start" || text === "/info") { return removeAudioHandlers(bot, chatId) }
+      if (ctx.chat.id === chatId) {
+        let normalizedFilename
+        let audioFile
+        let videoTitle
+        const videoUrl = ctx.message.text;
+        const normalVideoUrl = await checkYoutubeVideoUrl(ctx, chatId, videoUrl)
 
-      const videoUrl = ctx.message.text;
-      // console.log("videoUrl::", videoUrl)
-      const normalVideoUrl = await getCleanVideoUrl(ctx, chatId, videoUrl)
-      // console.log("normalVideoUrl:33:", normalVideoUrl)
-      // console.log("flags:33:", flags)
-
-
-      await youtubedl(normalVideoUrl, flags)
-        .then(async output => {
-          const authorName = output.uploader
-          videoTitle = output.title
-          let filename = `${authorName}-${videoTitle}`
-            .replace(/[/\\?%*:|"<>]/g, "")
-            .replace(/"/g, "'")
-            .substr(0, 64)
-          normalizedFilename = unorm.nfc(`${filename}.mp3`)
-          // console.log("normalizedFilename:99:", normalizedFilename)
-        })
-        .catch(error => {
-          console.log("error:", error)
-        })
-      // console.log("normalizedFilename:55:", normalizedFilename)
+        await youtubedl(normalVideoUrl, flags)
+          .then(async output => {
+            const authorName = output.uploader
+            videoTitle = output.title
+            let filename = `${authorName}-${videoTitle}`
+              .replace(/[/\\?%*:|"<>]/g, "")
+              .replace(/"/g, "'")
+              .substr(0, 64)
+            normalizedFilename = unorm.nfc(`${filename}.mp3`)
+          })
+          .catch(error => {
+            console.log("error:", error)
+          })
 
 
-      try {
-        audioFile = await AudioFile.findOne({ where: { videoLink: normalVideoUrl } })
-        // console.log("normalVideoUrl:66:", normalVideoUrl)
+        try {
+          audioFile = await AudioFile.findOne({ where: { videoLink: normalVideoUrl } })
+          if (!audioFile) {
+            console.log('Запись о файле не найдена в базе данных');
+            await downloadYoutubedl(ctx, chatId, botName, videoTitle, normalizedFilename, normalVideoUrl)
 
-        // console.log("audioFile::", audioFile.audioLink)
-        if (!audioFile) {
-          console.log('Запись о файле не найдена в базе данных');
+            return console.log(`Файл ${normalizedFilename} отправлен в чат ${chatId}`)
+
+          }
+
+          const telegramFile = await bot.telegram.getFile(audioFile.audioLink)
+
+          if (telegramFile) {
+            const fileSize = telegramFile.file_size
+            await sendAudioFromFileId(ctx, normalizedFilename, botName, audioFile.audioLink, fileSize)
+          }
+        } catch (error) {
+          await AudioFile.destroy({ where: { videoLink: normalVideoUrl } })
           await downloadYoutubedl(ctx, chatId, botName, videoTitle, normalizedFilename, normalVideoUrl)
-
-          return console.log(`Файл ${normalizedFilename} отправлен в чат ${chatId}`)
-
+          console.error('Ошибка при получении файла:', error)
         }
-
-        const telegramFile = await bot.telegram.getFile(audioFile.audioLink)
-
-        if (telegramFile) {
-          const fileSize = telegramFile.file_size
-          await sendAudioFromFileId(ctx, normalizedFilename, botName, audioFile.audioLink, fileSize)
-        }
-      } catch (error) {
-        await AudioFile.destroy({ where: { videoLink: normalVideoUrl } })
-        await downloadYoutubedl(ctx, chatId, botName, videoTitle, normalizedFilename, normalVideoUrl)
-        console.error('Ошибка при получении файла:', error)
       }
+
     } catch (error) {
       console.error('Произошла ошибка глобального try/catch (audioDownloader):', error)
-    } 
+    }
   })
 
 
@@ -89,123 +82,3 @@ module.exports = { audioDownloader }
 
 
 
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-// const unorm = require("unorm")
-// const { sendAudioFromFileId, downloadYoutubedl } = require("./audioFunction.js")
-// const { audioHandler, createAudioHandlers, removeAudioHandlers } = require("../../hendlers/audioHandler.js")
-// const { getCleanVideoUrl } = require("../../utils/checkUrl.js")
-// const { AudioFile } = require("../../models.js")
-// const youtubedl = require("youtube-dl-exec")
-
-
-// const flags = {
-//   dumpSingleJson: true,
-//   noCheckCertificates: true,
-//   noWarnings: true,
-//   preferFreeFormats: true,
-//   addHeader: ["referer:youtube.com", "user-agent:googlebot"]
-// }
-
-// async function audioDownloader(bot, chatId, botName) {
-//   let normalizedFilename
-//   let audioFile
-//   let videoTitle
-
-//   const messageAudioHandler = async ctx => {
-//     try {
-//       if (ctx.chat.id === chatId) {
-//         // проверяем, что сообщение от того же пользователя, которому отправили запрос
-//         const text = ctx.message.text
-//         if (text === "/start" || text === "/info") {
-//           return removeAudioHandlers(bot, chatId)
-//         }
-
-
-
-//         const videoUrl = ctx.message.text;
-//         // console.log("videoUrl::", videoUrl)
-//         const normalVideoUrl = await getCleanVideoUrl(ctx, chatId, videoUrl)
-//         // console.log("normalVideoUrl:33:", normalVideoUrl)
-//         // console.log("flags:33:", flags)
-
-
-
-
-//         await youtubedl(normalVideoUrl, flags)
-//           .then(async output => {
-//             const authorName = output.uploader
-//             videoTitle = output.title
-//             let filename = `${authorName}-${videoTitle}`
-//               .replace(/[/\\?%*:|"<>]/g, "")
-//               .replace(/"/g, "'")
-//               .substr(0, 64)
-//             normalizedFilename = unorm.nfc(`${filename}.mp3`)
-//             // console.log("normalizedFilename:99:", normalizedFilename)
-//           })
-//           .catch(error => {
-//             console.log("error:", error)
-//           })
-//         // console.log("normalizedFilename:55:", normalizedFilename)
-
-
-
-
-
-//         try {
-//           audioFile = await AudioFile.findOne({ where: { videoLink: normalVideoUrl } })
-//           // console.log("normalVideoUrl:66:", normalVideoUrl)
-
-//           // console.log("audioFile::", audioFile.audioLink)
-//           if (!audioFile) {
-//             console.log('Запись о файле не найдена в базе данных');
-//             await downloadYoutubedl(ctx, chatId, botName, videoTitle, normalizedFilename, normalVideoUrl)
-
-//             return console.log(`Файл ${normalizedFilename} отправлен в чат ${chatId}`)
-
-//           }
-
-//           const telegramFile = await bot.telegram.getFile(audioFile.audioLink)
-
-//           if (telegramFile) {
-//             const fileSize = telegramFile.file_size
-//             await sendAudioFromFileId(ctx, normalizedFilename, botName, audioFile.audioLink, fileSize)
-//           }
-//         } catch (error) {
-//           await AudioFile.destroy({ where: { videoLink: normalVideoUrl } })
-//           await downloadYoutubedl(ctx, chatId, botName, videoTitle, normalizedFilename, normalVideoUrl)
-//           console.error('Ошибка при получении файла:', error);
-//         }
-//       }
-//     } catch (error) {
-//       console.error('Произошла ошибка глобального try/catch (audioDownloader):', error);
-//     }
-//   }
-
-//   // if (!audioHandler.has(chatId)) {
-//   //   createAudioHandlers(bot, chatId, messageAudioHandler)
-//   // }
-// }
-
-
-// module.exports = { audioDownloader }
